+++ main.go
@@ -70,6 +70,15 @@ func main() {
 				Catch: app.catchPanic,
 			},
 		},
+		// https://github.com/wailsapp/wails/issues/3563
+		// on linux cause DomReady triggering
+		// on windows doesn't work at all
+		//DragAndDrop: &options.DragAndDrop{
+		//	EnableFileDrop: true,
+		//	DisableWebViewDrop: false,
+		//	CSSDropProperty:    "--drop-files",
+		//	CSSDropValue:       "drop",
+		//},
 		Windows: &windows.Options{
 			IsZoomControlEnabled: false,
 			Theme:                windows.Dark,